u64 -> u32

+++ testable-simd-models/src/abstractions/bitvec.rs
@@ -15,7 +15,7 @@ use std::fmt::Formatter;
 /// making the bit pattern more human-readable. The type also implements indexing,
 /// allowing for easy access to individual bits.
 #[derive(Copy, Clone, Eq, PartialEq)]
-pub struct BitVec<const N: u64>(FunArray<N, Bit>);
+pub struct BitVec<const N: u32>(FunArray<N, Bit>);
 
 /// Pretty prints a bit slice by group of 8
 fn bit_slice_to_string(bits: &[Bit]) -> String {
@@ -33,15 +33,15 @@ fn bit_slice_to_string(bits: &[Bit]) -> String {
         .into()
 }
 
-impl<const N: u64> core::fmt::Debug for BitVec<N> {
+impl<const N: u32> core::fmt::Debug for BitVec<N> {
     fn fmt(&self, f: &mut Formatter<'_>) -> Result<(), std::fmt::Error> {
         write!(f, "{}", bit_slice_to_string(&self.0.as_vec()))
     }
 }
 
-impl<const N: u64> core::ops::Index<u64> for BitVec<N> {
+impl<const N: u32> core::ops::Index<u32> for BitVec<N> {
     type Output = Bit;
-    fn index(&self, index: u64) -> &Self::Output {
+    fn index(&self, index: u32) -> &Self::Output {
         self.0.get(index)
     }
 }
@@ -75,19 +75,19 @@ fn int_from_bit_slice<T: TryFrom<i128> + MachineInteger + Copy>(bits: &[Bit]) ->
     };
     n
 }
-impl<const N: u64> BitVec<N> {
+impl<const N: u32> BitVec<N> {
     /// Constructor for BitVec. `BitVec::<N>::from_fn` constructs a bitvector out of a function that takes usizes smaller than `N` and produces bits.
-    pub fn from_fn<F: Fn(u64) -> Bit>(f: F) -> Self {
+    pub fn from_fn<F: Fn(u32) -> Bit>(f: F) -> Self {
         Self(FunArray::from_fn(f))
     }
     /// Convert a slice of machine integers where only the `d` least significant bits are relevant.
-    pub fn from_slice<T: Into<i128> + MachineInteger + Copy>(x: &[T], d: u64) -> Self {
+    pub fn from_slice<T: Into<i128> + MachineInteger + Copy>(x: &[T], d: u32) -> Self {
         Self::from_fn(|i| Bit::of_int::<T>(x[(i / d) as usize], (i % d) as u32))
     }
 
     /// Construct a BitVec out of a machine integer.
     pub fn from_int<T: Into<i128> + MachineInteger + Copy>(n: T) -> Self {
-        Self::from_slice::<T>(&[n], T::bits() as u64)
+        Self::from_slice::<T>(&[n], T::bits() as u32)
     }
 
     /// Convert a BitVec into a machine integer of type `T`.
@@ -115,12 +115,12 @@ impl<const N: u64> BitVec<N> {
     }
 }
 
-impl<const N: u64> BitVec<N> {
-    pub fn chunked_shift<const CHUNK: u64, const SHIFTS: u64>(
+impl<const N: u32> BitVec<N> {
+    pub fn chunked_shift<const CHUNK: u32, const SHIFTS: u32>(
         self,
         shl: FunArray<SHIFTS, i128>,
     ) -> BitVec<N> {
-        fn chunked_shift<const N: u64, const CHUNK: u64, const SHIFTS: u64>(
+        fn chunked_shift<const N: u32, const CHUNK: u32, const SHIFTS: u32>(
             bitvec: BitVec<N>,
             shl: FunArray<SHIFTS, i128>,
         ) -> BitVec<N> {
@@ -134,7 +134,7 @@ impl<const N: u64> BitVec<N> {
                 };
                 let local_index = (nth_bit as i128).wrapping_sub(shift);
                 if local_index < CHUNK as i128 && local_index >= 0 {
-                    let local_index = local_index as u64;
+                    let local_index = local_index as u32;
                     bitvec[nth_chunk * CHUNK + local_index]
                 } else {
                     Bit::Zero
